+++ src/meta/assessment/validation/nationalDataPoint/index.ts
@@ -1,0 +1,5 @@
+import { validate } from 'meta/assessment/validation/nationalDataPoint/validate'
+
+export const NationalDataPointValidator = {
+  validate,
+}
