+++ src/server/controller/cycleData/validations/ndp/validateNationalDataPoint.ts
@@ -1,9 +1,6 @@
-import { ODPs } from 'meta/assessment/odps'
 import { OriginalDataPoint } from 'meta/assessment/originalDataPoint'
+import { NationalDataPointValidator } from 'meta/assessment/validation/nationalDataPoint'
 import { NDPValidation } from 'meta/assessment/validation/ndp'
-import { Objects } from 'utils/objects'
-
-import { buildNationalClassValidation } from './buildNationalClassValidation'
 
 type Props = {
   nationalDataPoint: OriginalDataPoint
@@ -11,22 +8,6 @@ type Props = {
 
 export const validateNationalDataPoint = (props: Props): NDPValidation => {
   const { nationalDataPoint } = props
-  const validation: NDPValidation = {}
-
-  if (!ODPs.validateYear(nationalDataPoint)) {
-    validation.year = { valid: false }
-  }
-
-  nationalDataPoint.nationalClasses?.forEach((nationalClass, index) => {
-    const { placeHolder, uuid } = nationalClass
-    if (placeHolder || Objects.isEmpty(uuid)) return
-
-    const nationalClassValidation = buildNationalClassValidation({ nationalDataPoint, index })
-    if (Objects.isEmpty(nationalClassValidation)) return
-
-    validation.nationalClasses ??= {}
-    validation.nationalClasses[uuid] = nationalClassValidation
-  })
-
-  return validation
+  return NationalDataPointValidator.validate(nationalDataPoint)
+  // TODO: Validate national data point data sources
 }
